+++ src/types.js
@@ -29,7 +29,8 @@ export type Expiration = {
 /**
  * @typedef {Object} MicroMemoizeOptions
  *
- * @property {function} isEqual a custom equality comparator used for key matching
+ * @property {function} isEqual a custom equality comparator used for specific key argument matching
+ * @property {function} [isMatchingKey] a custom equality comparator used for full key matching
  * @property {boolean} isPromise is the return value a promise or not
  * @property {number} maxSize the maximum size of values to store in cache
  * @property {function} [onCacheAdd] method called when cache has been added to
@@ -39,6 +40,7 @@ export type Expiration = {
  */
 export type MicroMemoizeOptions = {
   isEqual: Function,
+  isMatchingKey: ?Function,
   isPromise: boolean,
   maxSize: number,
   onCacheAdd: ?Function,
@@ -51,19 +53,20 @@ export type MicroMemoizeOptions = {
  * @typedef {Object} Options
  *
  * @property {function} [equals] a custom equality comparator used for key matching
- * @property {boolean} isPromise is the return value a promise or not
- * @property {boolean} isSerialized should the parameters be serialized for the cache key
+ * @property {boolean} [isPromise] is the return value a promise or not
+ * @property {boolean} [isSerialized] should the parameters be serialized for the cache key
+ * @property {function} [matchesKey] the function to match full key equality
  * @property {number} [maxAge] the TTL for the return value to live in cache
  * @property {number} [maxArgs] the maximum number of arguments to use as the cache key
- * @property {number} maxSize the maximum size of values to store in cache
+ * @property {number} [maxSize] the maximum size of values to store in cache
  * @property {function} [onCacheAdd] method called when cache item is added
  * @property {function} [onCacheChange] method called when cache is added to or reordered
  * @property {function} [onCacheHit] method called when existing cache item is retrieved
  * @property {function} [onExpire] method to call when cache is expired
  * @property {string} [profileName] the name to use when profiling for statistics
- * @property {boolean} shouldSerializeFunctions should a custom replacer that includes functions be used in serialization
+ * @property {boolean} [shouldSerializeFunctions] should a custom replacer that includes functions be used in serialization
  * @property {function} [transformArgs] method to transform the arguments
- * @property {boolean} updateExpire should the expiration be updated when cache is hit
+ * @property {boolean} [updateExpire] should the expiration be updated when cache is hit
  */
 export type Options = {
   equals?: Function,
@@ -71,6 +74,7 @@ export type Options = {
   isPromise?: boolean,
   isReact?: boolean,
   isSerialized?: boolean,
+  matchesKey?: Function,
   maxAge?: number,
   maxArgs?: number,
   maxSize?: number,