+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # moize CHANGELOG
 
+## 5.1.0
+
+* Add [`matchesKey`](#matcheskey) option
+* Change `isSerialized` to use `matchesKey` with custom `isMatchingKey` instead of standard `isEqual`
+
 ## 5.0.1
 
 * Include `fast-equals` as an explicit dependency